service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.34
numpy==2.0.1
pillow==10.4.0
torch
torchvision
python-multipart>=0.0.9